Initial work on vue component.
